requirements.txt.txt
Flask==2.2.3
python-dotenv==1.0.0
google-generativeai==0.8.4
